BC-1211 Move api-tests for console
