+++ apps/server/src/console/api-test/server-console.api.spec.ts
@@ -3,7 +3,7 @@ import { INestApplicationContext } from '@nestjs/common';
 import { BootstrapConsole, ConsoleService } from 'nestjs-console';
 import { ServerConsoleModule } from '@src/console/console.module';
 import { ConsoleWriterService } from '@shared/infra/console';
-import { execute, TestBootstrapConsole } from './bootstrap.console';
+import { execute, TestBootstrapConsole } from './test-bootstrap.console';
 
 describe('ServerConsole (API)', () => {
 	let app: INestApplicationContext;
